Move quick actions UI into lib
